refactor: split into smaller components
