+++ src/App.tsx
@@ -1,111 +1,12 @@
-import { useState, useMemo } from 'react';
-import { JsonForms } from '@jsonforms/react';
-import Grid from '@mui/material/Grid';
-import Button from '@mui/material/Button';
-import Typography from '@mui/material/Typography';
 import './App.css';
-import schema from './schema.json';
-import uischema from './uischema.json';
-import {
-  materialCells,
-  materialRenderers,
-} from '@jsonforms/material-renderers';
-import RatingControl from './components/RatingControl';
-import ratingControlTester from './ratingControlTester';
-
-const classes = {
-  container: {
-    padding: '1em',
-    width: '100%',
-  },
-  title: {
-    textAlign: 'center',
-    padding: '0.25em',
-  },
-  dataContent: {
-    display: 'flex',
-    justifyContent: 'center',
-    borderRadius: '0.25em',
-    backgroundColor: '#cecece',
-    marginBottom: '1rem',
-  },
-  resetButton: {
-    margin: 'auto !important',
-    display: 'block !important',
-  },
-  demoform: {
-    margin: 'auto',
-    padding: '1rem',
-  },
-};
-
-const initialData = {
-  name: 'Send email to Adrian',
-  description: 'Confirm if you have passed the subject\nHereby ...',
-  done: true,
-  recurrence: 'Daily',
-  rating: 3,
-};
-
-const renderers = [
-  ...materialRenderers,
-  //register custom renderers
-  { tester: ratingControlTester, renderer: RatingControl },
-];
+import { Header } from './components/Header';
+import { JsonFormsDemo } from './components/JsonFormsDemo';
 
 const App = () => {
-  const [data, setData] = useState<object>(initialData);
-  const stringifiedData = useMemo(() => JSON.stringify(data, null, 2), [data]);
-
-  const clearData = () => {
-    setData({});
-  };
-
   return (
     <>
-      <div className='App'>
-        <header className='App-header'>
-          <img src='./logo.svg' className='App-logo' alt='logo' />
-          <h1 className='App-title'>Welcome to JSON Forms with React</h1>
-          <p className='App-intro'>More Forms. Less Code.</p>
-        </header>
-      </div>
-
-      <Grid
-        container
-        justifyContent={'center'}
-        spacing={1}
-        style={classes.container}
-      >
-        <Grid item sm={6}>
-          <Typography variant={'h4'}>Bound data</Typography>
-          <div style={classes.dataContent}>
-            <pre id='boundData'>{stringifiedData}</pre>
-          </div>
-          <Button
-            style={classes.resetButton}
-            onClick={clearData}
-            color='primary'
-            variant='contained'
-            data-testid='clear-data'
-          >
-            Clear data
-          </Button>
-        </Grid>
-        <Grid item sm={6}>
-          <Typography variant={'h4'}>Rendered form</Typography>
-          <div style={classes.demoform}>
-            <JsonForms
-              schema={schema}
-              uischema={uischema}
-              data={data}
-              renderers={renderers}
-              cells={materialCells}
-              onChange={({ data }) => setData(data)}
-            />
-          </div>
-        </Grid>
-      </Grid>
+      <Header />
+      <JsonFormsDemo />
     </>
   );
 };
