+++ src/components/Header.tsx
@@ -1,0 +1,11 @@
+import { FC } from 'react';
+
+export const Header: FC = () => (
+  <div className='App'>
+    <header className='App-header'>
+      <img src='./logo.svg' className='App-logo' alt='logo' />
+      <h1 className='App-title'>Welcome to JSON Forms with React</h1>
+      <p className='App-intro'>More Forms. Less Code.</p>
+    </header>
+  </div>
+);
